Feature gate the `type_name` function (#2)
As discussed in #1, this is probably overly cautious. However, the costs
on our side is extremely low, and the cost of a breaking release is
sufficiently high, that I believe this feature gate is worth having.

To clarify why someone could want to disable this feature - in certain
cases (especially on the web), binary size is extremely important. If
the optimiser/linker doesn't notice that this function is never used
(which is especially likely given that this type is used for `dyn`
values), each vtable for `dyn AnyDebug` will need an implementation of
this. These implementations individually don't have a huge cost; a
single function and a (relatively) short string. However, we expect
there to be quite a few of these (probably about 1 per widget, for
example).

---------

Co-authored-by: Tom Churchman <thomas@churchman.nl>